+++ client/src/api/schemas/index.ts
@@ -37,7 +37,6 @@ export * from './getConfigurationsResponse';
 export * from './getEventsParams';
 export * from './hTTPValidationError';
 export * from './includedCondition';
-export * from './independentTestUploadResponse';
 export * from './lockedByUser';
 export * from './notificationKeys';
 export * from './notificationsToRender';
@@ -46,6 +45,7 @@ export * from './release';
 export * from './releaseNotes';
 export * from './releasesResponse';
 export * from './sectionUpdateInput';
+export * from './simulatorUploadResponse';
 export * from './updateCustomCodeInput';
 export * from './updateUserNotificationsRequest';
 export * from './uploadCustomCodesCsvInput';
