+++ refiner/app/api/v1/v1_router.py
@@ -3,15 +3,15 @@ from fastapi import APIRouter
 from .app_notifications import router as app_notifications_router
 from .conditions import router as conditions_router
 from .configurations.router import router as configurations_router
-from .demo import router as demo_router
 from .events import router as events_router
 from .info import router as info_router
 from .releases import router as releases_router
+from .simulator import router as simulator_router
 
 router = APIRouter(prefix="/v1")
 router.include_router(conditions_router)
 router.include_router(configurations_router)
-router.include_router(demo_router)
+router.include_router(simulator_router)
 router.include_router(events_router)
 router.include_router(info_router)
 router.include_router(releases_router)
