+++ refiner/tests/integration/test_demo.py
@@ -5,7 +5,7 @@ from zipfile import ZipFile
 import pytest
 from fastapi import status
 
-api_route_base = "/api/v1/demo"
+api_route_base = "/api/v1/simulator"
 
 
 @pytest.mark.asyncio
@@ -18,7 +18,7 @@ async def test_demo_upload_smoke(
     activate_config,
 ) -> None:
     """
-    Smoke test for the /api/v1/demo/upload endpoint.
+    Smoke test for the /api/v1/simulator/upload endpoint.
     Verifies that the endpoint processes a demo ZIP file and returns a 200 with expected top-level fields.
     """
