+++ refiner/tests/integration/test_ecr_refiner.py
@@ -44,7 +44,7 @@ async def test_zip_upload_covid_influenza_v1_1(
     activate_config,
 ):
     """
-    Integration test for /api/v1/demo/upload using Mon Mothma COVID+Influenza v1.1.
+    Integration test for /api/v1/simulator/upload using Mon Mothma COVID+Influenza v1.1.
     """
 
     test_name = "test_zip_upload_covid_influenza_v1_1"
@@ -62,7 +62,7 @@ async def test_zip_upload_covid_influenza_v1_1(
             "uploaded_file": (covid_influenza_v1_1_zip_path.name, f, "application/zip")
         }
         response = await authed_client.post(
-            "/api/v1/demo/upload",
+            "/api/v1/simulator/upload",
             files=files,
         )
 
@@ -105,7 +105,7 @@ async def test_zip_upload_zika_v3_1_1(
     activate_config,
 ):
     """
-    Integration test for /api/v1/demo/upload using Mon Mothma Zika v3.1.1.
+    Integration test for /api/v1/simulator/upload using Mon Mothma Zika v3.1.1.
     """
 
     test_name = "test_zip_upload_zika_v3_1_1"
@@ -117,7 +117,7 @@ async def test_zip_upload_zika_v3_1_1(
     with open(zika_v3_1_1_zip_path, "rb") as f:
         files = {"uploaded_file": (zika_v3_1_1_zip_path.name, f, "application/zip")}
         response = await authed_client.post(
-            "/api/v1/demo/upload",
+            "/api/v1/simulator/upload",
             files=files,
         )
